Addressed Code Review
The weird .setScoreLevel calls were from my original thought to switch between showing the score level and showing the pickup object, which would change when spitting out and when intaking. Forgot to remove them.
Made hasObject pattern have a 0.5 second expire, since we only need to see it once to know we have picked something up and the drivers would like to know where they are set to score